Create history page, added widgets (history_point.dart, task_model.dart), Biggest widgets are divided into small ones

+++ lib/Models/task_model.dart
@@ -1,6 +1,6 @@
 class TaskModel{
   final int id;
-  final double count;
+  final int count;
   final String type;
 
   TaskModel(this.id, this.count, this.type);
